ts-gcp-serverless-pubsub-bigtable/backend-app/requirements.txt
google-cloud-bigtable>=2.17.0